+++ django_mobile/context_processors.py
@@ -4,17 +4,8 @@ from django_mobile.conf import settings
 
 def flavour(request):
     return {
-        'flavour': get_flavour(),
-    }
-
-
-def platform(request):
-    return {
-        'client_platform': get_platform(request),
-    }
-
-
-def is_mobile(request):
-    return {
-        'is_mobile': get_flavour() == settings.DEFAULT_MOBILE_FLAVOUR,
+        'flavour': get_flavour(request),
+        'client_platform':get_platform(request),
+        'is_mobile': get_flavour(request) == settings.DEFAULT_MOBILE_FLAVOUR,
+        'STATIC_URL_MOBILE': settings.STATIC_URL_MOBILE,
     }